ecommerce-protocol/requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.7.3
rsa==4.0
six==1.12.0
